(PC-17135)[PRO] refactor: rename VenueProviderMangerV2 components
