+++ pro/src/pages/Offerers/Offerer/VenueV1/VenueEdition/VenueProvidersManager/AllocineProviderParamaters/AllocineProviderParameters.tsx
@@ -4,11 +4,11 @@ import { api } from 'apiClient/api'
 import { getError, isErrorAPIError } from 'apiClient/helpers'
 import { VenueProviderResponse } from 'apiClient/v1'
 import useNotification from 'hooks/useNotification'
-import { Button } from 'ui-kit'
 import { ButtonVariant } from 'ui-kit/Button/types'
+import { Button } from 'ui-kit/index'
 
-import AllocineProviderFormDialog from '../../VenueProvidersManager/AllocineProviderFormDialog/AllocineProviderFormDialog'
-import { getRequestErrorStringFromErrors } from '../../VenueProvidersManager/utils/getRequestErrorStringFromErrors'
+import AllocineProviderFormDialog from '../AllocineProviderFormDialog/AllocineProviderFormDialog'
+import { getRequestErrorStringFromErrors } from '../utils/getRequestErrorStringFromErrors'
 
 import style from './AllocineProviderParameters.module.scss'
 import { IAllocineProviderParametersValues } from './types'
