+++ pro/src/pages/Offerers/Offerer/VenueV1/VenueEdition/VenueProvidersManager/VenueProviderItem/VenueProviderItem.module.scss
