--- pro/src/pages/Offerers/Offerer/VenueV1/VenueEdition/VenueProvidersManager/VenueProviderItem/VenueProviderItem.scss
@@ -1,110 +1,0 @@
-@use "styles/variables/_colors.scss" as colors;
-@use "styles/mixins/_rem.scss" as rem;
-@use "styles/mixins/_fonts.scss" as fonts;
-
-.venue-provider-row {
-  align-items: center;
-  display: flex;
-  flex-direction: column;
-  width: 100%;
-
-  .venue-provider-item-info {
-    display: flex;
-    flex-direction: row;
-    width: 100%;
-  }
-
-  .provider-name-container {
-    @include fonts.caption;
-
-    align-content: center;
-    display: flex;
-    flex: 2;
-    flex-direction: column;
-    justify-content: center;
-    margin-left: rem.torem(20px);
-    text-transform: uppercase;
-  }
-
-  .venue-id-at-offer-provider-container-with-message {
-    @include fonts.caption;
-
-    display: inline;
-    flex: 4;
-    margin-left: rem.torem(20px);
-  }
-
-  .venue-id-at-offer-provider-container {
-    @include fonts.caption;
-
-    align-items: center;
-    display: flex;
-    flex: 4;
-    flex-direction: row;
-
-    .venue-id-at-offer-provider {
-      @include fonts.caption;
-
-      flex: 1.5;
-    }
-
-    .offers-container-counter {
-      flex: 1;
-      justify-content: flex-end;
-      padding-top: rem.torem(8px);
-
-      .number-of-offers-label {
-        display: inline-block;
-        padding-left: rem.torem(5px);
-        vertical-align: top;
-      }
-
-      .offers-counter-icon {
-        width: rem.torem(22px);
-        height: rem.torem(22px);
-        fill: colors.$primary;
-      }
-    }
-  }
-
-  .offers-container-message {
-    flex: 1.5;
-    justify-content: flex-end;
-    padding-top: rem.torem(8px);
-  }
-
-  .venue-informations {
-    align-self: flex-start;
-
-    &-sync-item {
-      align-items: center;
-      color: colors.$grey-dark;
-      display: inline-flex;
-
-      &::before {
-        background-color: colors.$green-valid;
-        border-radius: 50%;
-        content: "";
-        display: inline-block;
-        height: rem.torem(8px);
-        margin-right: rem.torem(8px);
-        width: rem.torem(8px);
-      }
-
-      @include fonts.caption;
-    }
-
-    li {
-      margin-top: rem.torem(8px);
-
-      :first-child {
-        color: colors.$grey-dark;
-      }
-    }
-
-    button {
-      margin-top: rem.torem(32px);
-      padding: 0 rem.torem(32px);
-    }
-  }
-}
