+++ pro/src/pages/Offerers/Offerer/VenueV1/VenueEdition/VenueProvidersManager/index.js
@@ -1,0 +1,1 @@
+export { default } from './VenueProvidersManager'
