--- pro/src/pages/Offerers/Offerer/VenueV1/VenueEdition/VenueProvidersManagerV2/index.js
@@ -1,1 +1,0 @@
-export { default } from './VenueProvidersManagerV2'
